extract header components
