+++ frontend/fingertips-frontend/components/organisms/Heatmap/components/header/areaHeader.styles.tsx
@@ -1,0 +1,39 @@
+import { GovukColours } from '@/lib/styleHelpers/colours';
+import { H4, Table } from 'govuk-react';
+import styled from 'styled-components';
+import { heatmapDataColumnWidth } from '../../heatmapUtil';
+
+export const StyledDiv = styled.div({
+  position: 'relative',
+  height: '50px',
+});
+
+export const StyledH4AreaScaled = styled(H4)({
+  textOverflow: 'ellipsis',
+  whiteSpace: 'nowrap',
+  overflowX: 'hidden',
+  display: 'block',
+  margin: '0px',
+  position: 'absolute',
+  top: 0,
+  left: 0,
+  padding: '4px 8px',
+  transformOrigin: 'top left',
+  transform: 'translate(0, 15px) rotate(-60deg)',
+});
+
+export const StyledH4PrimaryBenchmarkHeader = styled(StyledH4AreaScaled)({
+  backgroundColor: GovukColours.MidGrey,
+});
+
+export const StyledH4SecondaryBenchmarkHeader = styled(StyledH4AreaScaled)({
+  backgroundColor: GovukColours.LightGrey,
+});
+
+export const StyledCellHeaderArea = styled(Table.CellHeader)({
+  verticalAlign: 'bottom',
+  width: `${heatmapDataColumnWidth}px`,
+  paddingRight: '0px',
+  paddingLeft: '10px',
+  height: '300px',
+});
