+++ frontend/fingertips-frontend/components/organisms/Heatmap/components/header/areaSecondaryBenchmarkHeader.tsx
@@ -1,0 +1,23 @@
+import {
+  StyledCellHeaderArea,
+  StyledDiv,
+  StyledH4SecondaryBenchmarkHeader,
+} from './areaHeader.styles';
+
+interface AreaSecondaryBenchmarkHeaderProps {
+  content: string;
+}
+
+export const AreaSecondaryBenchmarkHeader: React.FC<
+  AreaSecondaryBenchmarkHeaderProps
+> = ({ content }: AreaSecondaryBenchmarkHeaderProps) => {
+  return (
+    <StyledCellHeaderArea>
+      <StyledDiv>
+        <StyledH4SecondaryBenchmarkHeader>
+          {content}
+        </StyledH4SecondaryBenchmarkHeader>
+      </StyledDiv>
+    </StyledCellHeaderArea>
+  );
+};
